Move blame tooltip content to its own component
GutterItem was becoming too large, new logic needs to go in there for
the age tooltip so moved the blame tooltip into a separate component
to keep things manageable.
